+++ Pod/Classes/Terminals/Utility/HpsTerminalEnums.m
@@ -142,7 +142,7 @@ NSString *const PAX_DEVICE_VERSION = @"1.35";
 		case ARQC:
 			return @"ARQC";
 }
-	[NSException raise:NSInvalidArgumentException format:@"The given crypto type, %ld, is not known.", cryptoType];
+	[NSException raise:NSInvalidArgumentException format:@"The given crypto type, %ld, is not known.", (long)cryptoType];
 	return nil; // Keep the compiler happy - does not understand above line never returns!
 }
 @end
